Refactor: serverless an services utils
